requirements.txt
annotated-types==0.7.0
asttokens==2.4.0
backcall==0.2.0
beautifulsoup4==4.13.5
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==1.2.0
Flask==3.1.2
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
httplib2==0.30.0
idna==3.10
ipykernel==6.25.2
ipython==8.15.0
itsdangerous==2.2.0
jedi==0.19.0
Jinja2==3.1.6
jupyter_client==8.3.1
jupyter_core==5.3.1
MarkupSafe==3.0.2
matplotlib==3.8.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
packaging==23.1
pandas==2.1.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.16.1
pyparsing==3.2.3
python-dateutil==2.8.2
python-dotenv==1.1.1
pywin32==306
pyzmq==25.1.1
requests==2.32.5
rsa==4.9.1
scikit-learn==1.3.1
seaborn==0.12.2
six==1.16.0
soupsieve==2.8
stack-data==0.6.2
tornado==6.3.3
tqdm==4.67.1
traitlets==5.10.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.6
Werkzeug==3.1.3
